refactor: domain 폴더명을 model로 수정
