+++ src/model/LottoCalculator.js
