+++ src/model/LottoMachine.js
