+++ src/model/WinningLotto.js
